Move test_gpu_metrics test
gen_cluster doesn't cleanly handle pytest.importorskip

```
Task exception was never retrieved

future: <Task finished coro=<test_gpu_metrics() done, defined at
/home/travis/build/dask/distributed/distributed/tests/test_worker.py:1575>
exception=could not import 'pynvml': No module named 'pynvml'>

Traceback (most recent call last):

  File
  "/home/travis/miniconda/envs/test-environment/lib/python3.6/site-packages/_pytest/runner.py",
  line 244, in from_call
```

This moves the importorskip outside the test.
