+++ distributed/tests/test_worker.py
@@ -1572,18 +1572,6 @@ async def test_lifetime_stagger(c, s, a, b):
     assert 8 <= b.lifetime <= 12
 
 
-@gen_cluster()
-async def test_gpu_metrics(s, a, b):
-    pytest.importorskip("pynvml")
-    from distributed.diagnostics.nvml import count
-
-    assert "gpu" in a.metrics
-    assert len(s.workers[a.address].metrics["gpu"]["memory-used"]) == count
-
-    assert "gpu" in a.startup_information
-    assert len(s.workers[a.address].extra["gpu"]["name"]) == count
-
-
 @pytest.mark.asyncio
 async def test_bad_metrics(cleanup):
     def bad_metric(w):
